refactor: 상품 도메인 패키지 이동
