+++ src/main/java/shop/tryit/domain/item/entity/Category.java
@@ -1,7 +1,6 @@
-package shop.tryit.domain.item;
+package shop.tryit.domain.item.entity;
 
 public enum Category {
 
     CAT, DOG
-
 }
