+++ src/main/java/shop/tryit/domain/item/entity/Image.java
@@ -14,7 +14,6 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.tryit.domain.item.ImageType;
 
 @Entity
 @Getter
