+++ src/main/java/shop/tryit/domain/item/entity/ImageType.java
@@ -1,7 +1,6 @@
-package shop.tryit.domain.item;
+package shop.tryit.domain.item.entity;
 
 public enum ImageType {
 
     MAIN, DETAIL
-
 }
