+++ src/main/java/shop/tryit/repository/item/ItemCustom.java
@@ -2,7 +2,7 @@ package shop.tryit.repository.item;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.tryit.domain.item.ItemSearchCondition;
+import shop.tryit.domain.item.dto.ItemSearchCondition;
 import shop.tryit.domain.item.dto.ItemSearchDto;
 
 public interface ItemCustom {
