+++ src/main/java/shop/tryit/repository/item/ItemCustomImpl.java
@@ -18,9 +18,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
-import shop.tryit.domain.item.Category;
-import shop.tryit.domain.item.ImageType;
-import shop.tryit.domain.item.ItemSearchCondition;
+import shop.tryit.domain.item.entity.Category;
+import shop.tryit.domain.item.entity.ImageType;
+import shop.tryit.domain.item.dto.ItemSearchCondition;
 import shop.tryit.domain.item.dto.ItemSearchDto;
 import shop.tryit.domain.item.dto.QItemSearchDto;
 import shop.tryit.domain.item.entity.Image;
