+++ src/main/java/shop/tryit/web/item/ItemController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import shop.tryit.domain.common.Pages;
-import shop.tryit.domain.item.Category;
-import shop.tryit.domain.item.ItemSearchCondition;
+import shop.tryit.domain.item.entity.Category;
+import shop.tryit.domain.item.dto.ItemSearchCondition;
 import shop.tryit.domain.item.dto.ItemDto;
 import shop.tryit.domain.item.dto.ItemFormDto;
 import shop.tryit.domain.item.dto.ItemSearchDto;
